[ssd][refactoring] IOManager refatoring stage 3 - NAND IO
