+++ SSDSimulator/SSDSimulator/IOManager.h
@@ -1,15 +1,13 @@
 #pragma once
 #include "outputFile.h"
 #include "bufferFile.h"
-
-struct LbaEntry {
-    uint32_t address;
-    uint32_t data;
-};
+#include "nandFile.h"
 
 class IOManager {
 public:
-    IOManager(uint32_t maxLbaOfDevice) : maxLbaOfDevice(maxLbaOfDevice) {}
+    IOManager(uint32_t maxLbaOfDevice)  {
+        nand().setDeviceMaxLba(maxLbaOfDevice);
+    }
 
     OutputFile& output() { return outputFile; }
 #if OUTPUT_REFACTORING
@@ -26,32 +24,18 @@ public:
     std::vector<std::string> getBufferFileList();
 #endif
 
+    NandFile& nand() { return nandFile; }
+#if NAND_REFACTORING 
     void CheckAndCreateNandDataFile();
     void ProgramAllDatasToNand(const std::vector<LbaEntry>& lbaTable);
     void ReadAllDatasToInternalBuffer(std::vector<LbaEntry>& lbaTable);
     bool SplitStringToAddressAndData(std::string& line, LbaEntry* splitDatas);
     void CreateNewTempNandFileAndInitForTest();
     void deleteFileIfExists();
+#endif
 
 private:
     OutputFile outputFile;
     BufferFile bufferFile;
-
-    bool nandDataFileExist();
-    void FillZeroDataToAllAddresses(std::ofstream& nandDataFile);
-    void FillChangeDatasToAllAddresses(std::ofstream& nandDataFile,
-        const std::vector<LbaEntry>& lbaTable);
-    bool SuccessConvertToUINT(LbaEntry* splitDatas, std::string& addrStr,
-        std::string& dataStr);
-
-    const uint32_t maxLbaOfDevice;
-    const static uint32_t INIT_NAND_DATA = 0;
-    const std::string NAND_DATA_FILE = "ssd_nand.txt";
-    const std::string SEPARATOR = ";";
-
-#if BUFFER_REFACTORING
-    static const uint32_t NUM_OF_BUFFERS = 5;
-    const std::string BUFFER_FOLDER = "./buffer/";
-    const std::string EMPTY_BUFFER_FILE_SUFFIX = "_empty";
-#endif
+    NandFile nandFile;
 };
