+++ SSDSimulator/SSDSimulator/SSDSimulator.vcxproj
@@ -149,6 +149,7 @@
     <ClInclude Include="cmdBuffer.h" />
     <ClInclude Include="fileIO.h" />
     <ClInclude Include="IOManager.h" />
+    <ClInclude Include="nandFile.h" />
     <ClInclude Include="outputFile.h" />
     <ClInclude Include="ssdCmdErase.h" />
     <ClInclude Include="ssdCmdError.h" />
